requirement.txt
pyodbc==4.0.39
python-dotenv==1.0.0
pandas==1.5.3
openai==0.28.1
jupyter==1.0.0
requests==2.31.0
azure-core==1.29.2
azure-ai-ml
azure-core
azure-identity
azure-keyvault
azure-kusto-data
azure-search-documents==11.4.0b11
azure-storage-blob
pyodbc
bs4
jinja2
jsbeautifier
langchain
logger
matplotlib
openai
pandas
plotly
promptflow>=0.1.0b8
promptflow-tools
python-dotenv
ipywidgets
scikit-learn
scipy
tiktoken
tenacity
tqdm

types-PyYAML
types-requests
